Rename `Place.load_csc_db!` to `Place.load_db!`
`csc` is an abbreviation for the countries-states-cities-database project. The
struct this function returns used to be called `%Place.CSCDB{}`, but it was
renamed to `%Place.DB{}`, since the dataset is kind of an implementation detail
and its name should probably not be exposed in the public API. So it makes sense
that this function follows suit.